requirements.txt
# Speech recognition
speechrecognition>=3.10.0
pyaudio>=0.2.13
openai-whisper>=20231117

# NLU and Agent
openai>=1.3.0
langchain>=0.0.335

# MCP integration
requests>=2.31.0
websockets>=11.0.3

# Emotional TTS
elevenlabs
boto3>=1.28.0  # For Amazon Polly

# Utilities
numpy>=1.24.0
pydantic>=2.4.0
python-dotenv>=1.0.0

